requirements.txt
python-telegram-bot>=8
python-valve>=0.2
pyyaml
dataset
